requirements.txt
fastapi
uvicorn
pydantic
sympy
numpy
groq
python-multipart
google-generativeai
pillow
httpx
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
supabase
easyocr
pix2tex
